+++ velox/common/memory/ByteStream.cpp
@@ -97,7 +97,7 @@ void ByteInputStream::seekp(std::streampos position) {
     }
     toSkip -= range.size;
   }
-  VELOX_FAIL("Seeking past end of ByteInputStream: {}", position);
+  VELOX_FAIL("Seeking past end of ByteInputStream: {}", static_cast<int>(position));
 }
 
 void ByteInputStream::next(bool throwIfPastEnd) {
@@ -283,7 +283,7 @@ void ByteOutputStream::seekp(std::streampos position) {
     }
     toSkip -= range.size;
   }
-  VELOX_FAIL("Seeking past end of ByteOutputStream: {}", position);
+  VELOX_FAIL("Seeking past end of ByteOutputStream: {}", static_cast<int>(position));
 }
 
 void ByteOutputStream::flush(OutputStream* out) {
